Fixes #15 - Scroll element into view first
Before getting the location of the element, capturing the screenshot, and saving; make sure the element is in view.  Especially in chrome, the screenshot is constrained to the viewport, so an element may not be in the saved screenshot if it is outside of the viewport.  Use getLocationInView to first move the element into view.